+++ frontend/src/hooks/useUpdateStationChargerReport.ts
@@ -1,0 +1,39 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { getLocalStorage } from '@utils/storage';
+
+import { modalActions } from '@stores/modalStore';
+
+import { BASE_URL, LOCAL_KEY_TOKEN } from '@constants';
+
+const fetchReportCharger = async (stationId: number) => {
+  const token = getLocalStorage<number>(LOCAL_KEY_TOKEN, -1);
+  return fetch(`${BASE_URL}/stations/${stationId}/reports`, {
+    method: 'POST',
+    body: JSON.stringify({ stationId }),
+    headers: {
+      Authorization: `Token ${token}`,
+      'Content-Type': 'application/json',
+    },
+  });
+};
+export const useUpdateStationChargerReport = () => {
+  const queryClient = useQueryClient();
+
+  const { mutate } = useMutation({
+    mutationFn: fetchReportCharger,
+    onSuccess: () => {
+      alert('신고가 완료됐습니다.');
+      modalActions.closeModal();
+    },
+    onSettled: () => {
+      queryClient.invalidateQueries({ queryKey: ['isStationChargerReported'] });
+    },
+  });
+
+  const updateStationChargerReport = (stationId: number) => {
+    mutate(stationId);
+  };
+
+  return { updateStationChargerReport };
+};
